Fixes area totals in certain situations (#29)
We need a new entry with all chunks in the response from `getPushDelta`,
so that we can correctly calculate the per-area totals in `comments.js`.

Previously, we were using `chunks`, which only includes the chunks which
changed in size. This would mean that if a chunk were to no longer be
used in one area, but remain unchanged, the total for that area would be
incorrectly calculated.